refactor(rpc-usage): mergeRpcUsageWindows lives in the chain model layer (review)
The merge is a pure model operation on RpcUsageWindow — moved next to
the type in packages/chain/src/rpc-usage.ts and exported from the
barrel alongside it (a deliberate widening of the earlier type-only
narrowing: it now has a genuine cross-package consumer and carries no
transport/parsing internals). publisher-runner and lifecycle import it
from dkg-chain; the daemon log module is back to formatting and
scheduling only. Merge-semantics unit tests moved to the chain suite;
the composite-source wiring test stays in cli. Suites: chain 8, cli 11,
agent 4 — all green.
