+++ packages/cli/test/rpc-usage-log.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import { formatRpcUsageLines, emitRpcUsage, mergeRpcUsageWindows, startRpcUsageTelemetry } from '../src/daemon/rpc-usage-log.js';
+import { mergeRpcUsageWindows } from '@origintrail-official/dkg-chain';
+import { formatRpcUsageLines, emitRpcUsage, startRpcUsageTelemetry } from '../src/daemon/rpc-usage-log.js';
 
 /**
  * The rpc_usage line format is a CONTRACT with the Grafana dashboards (parsed
@@ -42,26 +43,7 @@ describe('formatRpcUsageLines — the Grafana-facing rpc_usage contract', () => 
   });
 });
 
-describe('mergeRpcUsageWindows — one process, many trackers (agent + per-wallet publisher adapters)', () => {
-  it('sums per-method, total and lifetime across windows', () => {
-    const merged = mergeRpcUsageWindows(
-      { byMethod: { eth_call: 5, eth_estimateGas: 2 }, total: 7, lifetimeTotal: 100 },
-      { byMethod: { eth_call: 3, eth_sendRawTransaction: 4 }, total: 7, lifetimeTotal: 50 },
-    );
-    expect(merged).toEqual({
-      byMethod: { eth_call: 8, eth_estimateGas: 2, eth_sendRawTransaction: 4 },
-      total: 14,
-      lifetimeTotal: 150,
-    });
-  });
-
-  it('skips undefined inputs (adapters without the capability / runtime not started)', () => {
-    const w = { byMethod: { eth_call: 1 }, total: 1, lifetimeTotal: 1 };
-    expect(mergeRpcUsageWindows(undefined, w, undefined)).toEqual(w);
-    expect(mergeRpcUsageWindows(undefined, undefined)).toBeUndefined();
-    expect(mergeRpcUsageWindows()).toBeUndefined();
-  });
-
+describe('composite daemon source — agent + publisher-runtime windows merged at drain time', () => {
   it('COMPOSITE SOURCE end-to-end: publisher-runtime traffic reaches the emitted rpc_usage lines', () => {
     // The exact daemon wiring shape: the source merges the agent window with
     // the (lazily started) publisher runtime's window at every drain. The
